Divide person classes into different files
I did that according to code review option suggestion
