+++ nameable_decorators/base_decorator.rb
@@ -1,0 +1,13 @@
+# base decorator
+require_relative 'nameable'
+
+class BaseDecorator < Nameable
+  def initialize(nameable)
+    super()
+    @nameable = nameable
+  end
+
+  def correct_name
+    @nameable.correct_name
+  end
+end
